% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{easy_pop}
\alias{easy_pop}
\title{Easy Population}
\usage{
easy_pop(vector)
}
\description{
Function that displays all of the necesssary population data needed for a Quest/Homework
Accepts: data in vector format
}
